mark_3/requirements.txt
# Raspberry Pi 5 Offline AI Assistant Requirements
# Core dependencies for the AI assistant

# LLM and AI
llama-cpp-python==0.2.20
ultralytics==8.0.196

# Computer Vision and Camera
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# RAG and Embeddings (Optional)
sentence-transformers==2.2.2
chromadb==0.4.15

# System Monitoring
psutil==5.9.6

# Utilities
numpy==1.24.3
Pillow==10.0.1

# Audio and Voice
pyaudio==0.2.14
vosk==0.3.45
pyttsx3==2.90

# Hardware Interfaces
RPi.GPIO==0.7.1
adafruit-circuitpython-ssd1306==2.12.14
adafruit-blinka
#==8.19.1

# Additional dependencies for Raspberry Pi
# Note: Some packages may need to be installed via apt-get
# sudo apt-get install python3-opencv python3-numpy python3-pyaudio
# sudo apt-get install espeak espeak-data libespeak1 libespeak-dev
# sudo apt-get install i2c-tools python3-smbus
# sudo apt-get install portaudio19-dev python3-pyaudio (if pyaudio fails to install)

# Optional: For better performance
# torch==2.0.1  # Uncomment if you want to use PyTorch instead of ultralytics
# torchvision==0.15.2

# Development and debugging (optional)
# matplotlib==3.7.2
# jupyter==1.0.0
